requirements.txt
easyocr
opencv-python
numpy
pydantic
Pillow
